Bug Resolution: When a dice hits a side of the box, it sticks there… For example, hitting the bottom border. It does get deflected and the dice tries to move upwards.. However, since it is still overlapping the border, it gets deflected again so it moves downwards again. Thus, the dice sticks at this point. To solve that, I’m adding 4 booleans to allow the dice to hit a border just once.
Dice Container:
This is the container that will hold the dices within its boundaries. Box borders will be rectangles defined in Constants. Game controller will choose which container to use. Also, the container defines the X, Y margins between which dices randomly are placed on the screen (within the container)

Added sound effect when a dice is thrown